# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Sign in Page",
        "responses": {
          "200": {
            "description": "Login page retrieved successfully"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login with registered Email or Username and Password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/loginInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "You are logged in"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      }
    },
    "/auth/register": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Sign up Page",
        "responses": {
          "200": {
            "description": "Sign up page retrieved successfully"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Tp register enter FirstName, LastName, Username, Password, Address, Email and Telephone",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/signupInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created account"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Get all brands",
        "responses": {
          "200": {
            "description": "Successfully retrieved all brands"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Create a new brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Edit an existing brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editBrand"
            }
          },
          {
            "name": "BrandId",
            "in": "path",
            "description": "ID number of a specific brand",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/brands/{BrandId}": {
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "BrandId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific brand"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin-Brands"
        ],
        "description": "Get all brands",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all brands"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin-Brands"
        ],
        "description": "Create a new brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
    "loginInfo": {
      "type": "object",
      "properties": {
        "EmailorUsername": {
          "type": "string",
          "example": "'sample@sample.com'/'johnDoe'"
        },
        "Password": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "EmailorUsername",
        "Password"
      ]
    },
    "signupInfo": {
      "type": "object",
      "properties": {
        "FirstName": {
          "type": "string",
          "example": "John"
        },
        "LastName": {
          "type": "string",
          "example": "Doe"
        },
        "Username": {
          "type": "string",
          "example": "johnDoe"
        },
        "Password": {
          "type": "string",
          "example": "123456"
        },
        "Address": {
          "type": "string",
          "example": "somewhere, earth"
        },
        "Email": {
          "type": "string",
          "example": "sample@sample.com"
        },
        "Telephone": {
          "type": "string",
          "example": "00123456789"
        }
      },
      "required": [
        "FirstName",
        "LastName",
        "Username",
        "Password",
        "Address",
        "Email",
        "Telephone"
      ]
    },
    "createBrand": {
      "type": "object",
      "properties": {
        "Brand": {
          "type": "string",
          "example": "Example Brand"
        }
      },
      "required": [
        "Brand"
      ]
    },
    "editBrand": {
      "type": "object",
      "properties": {
        "NewBrand": {
          "type": "string",
          "example": "Edited Brand"
        }
      },
      "required": [
        "NewBrand"
      ]
    },
...
  }
}